add a server to handle avatar lookups
The service we were using to get twitter avatars client-side is no
longer working, so this is a very simple TypeScript server that does
exactly one thing, return the avatar for a twitter handle. It's very
small, untested, and doesn't handle errors well, but it should do the
trick and can stand as a basis for future server integration.